Created more folders for views for increased structure. Created custom element views (CustomButton, CustomTextField, CustomTitle, ForgotPasswordButton) for reusage. Added a background for the tab view for increased visibility on map.
