+++ BabyBrunch/Views/TabViews/MapView.swift
@@ -56,23 +56,11 @@ struct MapView: View {
                     secondaryButton: .cancel(Text("Cancel"))
                 )
             }
-            Button("Where am I?") {
-                vm.mapShouldBeUpdated = true
-                vm.checkIfLocationServicesEnabled()
-            }
-            
-            .foregroundColor(.white)
-            .frame(width: 150, height: 10)
-            .padding()
-            .background(Color("oldRose"))
-            .cornerRadius(10)
-            .overlay(
-                RoundedRectangle(cornerRadius: 10)
-                    .stroke(Color.black, lineWidth: 1)
-            )
-            .padding(.bottom, 600)
+           CustomButton(label: "Where am I?", backgroundColor: "oldRose", width: 150) {
+              vm.mapShouldBeUpdated = true
+              vm.checkIfLocationServicesEnabled()
+           }.padding(.bottom, 600)
         }
-  
     }
     
     // MARK: Functions
